Crate mtg_rust

The rust crate for the Mini Tardis Games addon of the Minecraft mod Mini Tardis.
Setup your Game using game_impl and you’re ready to go!

Macros

game_impl

Connects a type implementing Game with the underlying WASM functions.
Use this macro at the top level once with your game type as the argument.

Structs

Canvas

A canvas that can be manipulated via various methods and is available in draw callbacks.
Analogous to Map Canvas API’s DrawableCanvas class.

Screen

A (console) screen that contains methods that can be called from most callbacks.
Analogous to Mini Tardis’ ScreenBlockEntity class.

Enums

ClickType

Indicates the type of a Game::on_click callback, usually corresponding to a mouse button.
Analogous to the class with the same name in yarn mappings.

LogLevel

A level of importance for the log function.
Analogous to SLF4J’s Level enum.

SoundCategory

A category for a certain sound event, used by Screen::play_sound.
Analogous to the class with the same name in yarn mappings.

Traits

Game

The central game trait that your state type should implement.
Register that type using game_impl.

Functions

get_persistent_data

log

Logs a message.
Forwards to the mod’s SLF4J Logger’s log method.

nano_time

Returns an arbitrary nanosecond value to measure elapsed time between two calls.
Analogous to Java’s System.nanoTime() method.

save_persistent_data